Refactor type casting and improve code consistency across multiple files
- Replaced `static_cast<size_t>` with direct size type usage in various locations to enhance code clarity and reduce unnecessary type conversions.
- Updated memory management functions to utilize `std::uninitialized_copy_n` and `std::copy_n` for better performance and safety when handling trivially copyable types.
- Adjusted test cases to reflect changes in type handling, ensuring consistency and correctness in comparisons and data manipulations.

These updates contribute to a cleaner and more efficient codebase, improving maintainability and readability.

+++ app/src/log_model.cpp
@@ -715,7 +715,7 @@ void LogModel::SetRetentionCap(size_t cap)
     }
 
     const bool paused = mSink && mSink->IsPaused();
-    const size_t visible = static_cast<size_t>(mLogTable.RowCount());
+    const size_t visible = mLogTable.RowCount();
 
     if (!paused)
     {
